minor improvements and tests update
Passing ‘all’ to the integrations.all method returned a 404 status code
because https://github.com/integrations/features/all doesn’t exists.
This can be a bit misleading since all does  appear as a category in
the integrations page. Added a fallback conditional just in case ‘all’
is passed to the integrations.all method.